First round of fixes
